fix initial size/location of minimap.
Added an empty override of Window_MiniMap.PreOpen() to remove vanilla behaviour of centering and resizing window to InitialSize.

Removed redundant Window_MiniMap constructor that allowed setting size - size is always stored in minimapSize, and set from there in the base constructor. The only use of the deleted constructor took size from the minimapSize, and then fed it back in.

Removed unreachable error check if creation of window failed. If that does fail, the constructor would throw an exception before the error check could ever be reached.